Add an option for a Step to use geojson as a source of data rather than merely a parent Step.
This would be used to avoid having to have a parent Step for the paremeter needed.  A typical use case would be to have a Spatial Dataset step not have to rely on a Spatial Input step for source geometry.  Instead, geojson information can be passed to the geometry_source option for the Spatial Dataset step.  This would create the geometry from the geojson and have that appear for the Spatial Dataset step, rather than have the user enter some geometry in a parent Spatial Input step.

ex:

geojson_source = {'type': 'FeatureCollection', .....}

step = SpatialDatasetStep(
    name='Step',
    options={
        'geometry_source': {
            SpatialDatasetStep.OPT_GEOSON_SOURCE: geojson_source,
        }
    }
    'dataset_title': 'title',
    'template_dataset': ....
)